requirements.txt
einops==0.4.1
matplotlib==3.5.1
numpy==1.21.2
Pillow==9.1.1
seaborn==0.11.2
torch==1.11.0
torchtext==0.12.0
torchvision==0.12.0
tqdm==4.62.3
